fixed profile state
